requirements.txt
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.4.22
distro==1.9.0
Django==5.2.13
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.18.0
exceptiongroup==1.3.1
gunicorn==25.3.0
whitenoise==6.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
packaging==26.2
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
python-dotenv==1.2.2
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
